// frontend/src/api/generated/models/building-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BuildingDetail
 */
export interface BuildingDetail {
    /**
     * 
     * @type {number}
     * @memberof BuildingDetail
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'buildingCode': string;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'buildingName': string;
    /**
     * 
     * @type {number}
     * @memberof BuildingDetail
     */
    'warehouseId': number;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'warehouseCode': string;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'warehouseName'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BuildingDetail
     */
    'isActive': boolean;
    /**
     * 
     * @type {number}
     * @memberof BuildingDetail
     */
    'version': number;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof BuildingDetail
     */
    'updatedAt': string;
}